# platform-api/src/internal/handler/gateway_internal.go
/*
 *  Copyright (c) 2025, WSO2 LLC. (http://www.wso2.org) All Rights Reserved.
 *
 *  Licensed under the Apache License, Version 2.0 (the "License");
 *  you may not use this file except in compliance with the License.
 *  You may obtain a copy of the License at
 *
 *  http://www.apache.org/licenses/LICENSE-2.0
 *
 *  Unless required by applicable law or agreed to in writing, software
 *  distributed under the License is distributed on an "AS IS" BASIS,
 *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 *  See the License for the specific language governing permissions and
 *  limitations under the License.
 *
 */

package handler

import (
	"errors"
	"fmt"
	"log/slog"
	"net/http"
	"platform-api/src/internal/constants"
	"platform-api/src/internal/dto"
	"platform-api/src/internal/utils"

	"platform-api/src/internal/service"

	"github.com/gin-gonic/gin"
)

type GatewayInternalAPIHandler struct {
	gatewayService         *service.GatewayService
	gatewayInternalService *service.GatewayInternalAPIService
	slogger                *slog.Logger
}

func NewGatewayInternalAPIHandler(gatewayService *service.GatewayService,
	gatewayInternalService *service.GatewayInternalAPIService, slogger *slog.Logger) *GatewayInternalAPIHandler {
	return &GatewayInternalAPIHandler{
		gatewayService:         gatewayService,
		gatewayInternalService: gatewayInternalService,
		slogger:                slogger,
	}
}

// GetAPIsByOrganization handles GET /api/internal/v1/apis
func (h *GatewayInternalAPIHandler) GetAPIsByOrganization(c *gin.Context) {
	// Extract client IP for rate limiting
	clientIP := c.ClientIP()

	// Extract and validate API key from header
	apiKey := c.GetHeader("api-key")
	if apiKey == "" {
		h.slogger.Warn("Unauthorized access attempt - Missing API key", "clientIP", clientIP)
		c.JSON(http.StatusUnauthorized, utils.NewErrorResponse(401, "Unauthorized",
			"API key is required. Provide 'api-key' header."))
		return
	}

	// Authenticate gateway using API key
	gateway, err := h.gatewayService.VerifyToken(apiKey)
	if err != nil {
		h.slogger.Warn("Authentication failed", "clientIP", clientIP, "error", err)
		c.JSON(http.StatusUnauthorized, utils.NewErrorResponse(401, "Unauthorized",
			"Invalid or expired API key"))
		return
	}

	orgID := gateway.OrganizationID
	apis, err := h.gatewayInternalService.GetAPIsByOrganization(orgID)
	if err != nil {
		if errors.Is(err, constants.ErrOrganizationNotFound) {
			c.JSON(http.StatusNotFound, utils.NewErrorResponse(404, "Not Found",
				"Organization not found"))
			return
		}
		c.JSON(http.StatusInternalServerError, utils.NewErrorResponse(500, "Internal Server Error",
			"Failed to get apis"))
		return
	}

	// Create ZIP file from API YAML file
	zipData, err := utils.CreateAPIYamlZip(apis)
	if err != nil {
		h.slogger.Error("Failed to create ZIP file", "orgID", orgID, "error", err)
		c.JSON(http.StatusInternalServerError, utils.NewErrorResponse(500, "Internal Server Error",
			"Failed to create API package"))
		return
	}

	// Set headers for ZIP file download
	c.Header("Content-Type", "application/zip")
	c.Header("Content-Disposition", fmt.Sprintf("attachment; filename=\"apis-org-%s.zip\"", orgID))
	c.Header("Content-Length", fmt.Sprintf("%d", len(zipData)))

	// Return ZIP file
	c.Data(http.StatusOK, "application/zip", zipData)
}

// GetAPI handles GET /api/internal/v1/apis/:apiId
func (h *GatewayInternalAPIHandler) GetAPI(c *gin.Context) {
	// Extract client IP for rate limiting
	clientIP := c.ClientIP()

	// Extract and validate API key from header
	apiKey := c.GetHeader("api-key")
	if apiKey == "" {
		h.slogger.Warn("Unauthorized access attempt - Missing API key", "clientIP", clientIP)
		c.JSON(http.StatusUnauthorized, utils.NewErrorResponse(401, "Unauthorized",
			"API key is required. Provide 'api-key' header."))
		return
	}

	// Authenticate gateway using API key
	gateway, err := h.gatewayService.VerifyToken(apiKey)
	if err != nil {
		h.slogger.Warn("Authentication failed", "clientIP", clientIP, "error", err)
		c.JSON(http.StatusUnauthorized, utils.NewErrorResponse(401, "Unauthorized",
			"Invalid or expired API key"))
		return
	}

	orgID := gateway.OrganizationID
	gatewayID := gateway.ID
	apiID := c.Param("apiId")
	if apiID == "" {
		c.JSON(http.StatusBadRequest, utils.NewErrorResponse(400, "Bad Request",
			"API ID is required"))
		return
	}

	api, err := h.gatewayInternalService.GetActiveDeploymentByGateway(apiID, orgID, gatewayID)
	if err != nil {
		if errors.Is(err, constants.ErrDeploymentNotActive) {
			c.JSON(http.StatusNotFound, utils.NewErrorResponse(404, "Not Found",
				"No active deployment found for this API on this gateway"))
			return
		}
		if errors.Is(err, constants.ErrAPINotFound) {
			c.JSON(http.StatusNotFound, utils.NewErrorResponse(404, "Not Found",
				"API not found"))
			return
		}
		c.JSON(http.StatusInternalServerError, utils.NewErrorResponse(500, "Internal Server Error",
			"Failed to get API"))
		return
	}

	// Create ZIP file from API YAML file
	zipData, err := utils.CreateAPIYamlZip(api)
	if err != nil {
		h.slogger.Error("Failed to create ZIP file", "apiID", apiID, "error", err)
		c.JSON(http.StatusInternalServerError, utils.NewErrorResponse(500, "Internal Server Error",
			"Failed to create API package"))
		return
	}

	// Set headers for ZIP file download
	c.Header("Content-Type", "application/zip")
	c.Header("Content-Disposition", fmt.Sprintf("attachment; filename=\"api-%s.zip\"", apiID))
	c.Header("Content-Length", fmt.Sprintf("%d", len(zipData)))

	// Return ZIP file
	c.Data(http.StatusOK, "application/zip", zipData)
}

// CreateGatewayDeployment handles POST /api/internal/v1/apis/{apiId}/gateway-deployments
func (h *GatewayInternalAPIHandler) CreateGatewayDeployment(c *gin.Context) {
	// Extract client IP for logging
	clientIP := c.ClientIP()

	// Extract and validate API key from header
	apiKey := c.GetHeader("api-key")
	if apiKey == "" {
		h.slogger.Warn("Unauthorized access attempt - Missing API key", "clientIP", clientIP)
		c.JSON(http.StatusUnauthorized, utils.NewErrorResponse(401, "Unauthorized",
			"API key is required. Provide 'api-key' header."))
		return
	}

	// Authenticate gateway using API key
	gateway, err := h.gatewayService.VerifyToken(apiKey)
	if err != nil {
		h.slogger.Warn("Authentication failed", "clientIP", clientIP, "error", err)
		c.JSON(http.StatusUnauthorized, utils.NewErrorResponse(401, "Unauthorized",
			"Invalid or expired API key"))
		return
	}

	// Extract API ID from path parameter
	apiID := c.Param("apiId")
	if apiID == "" {
		c.JSON(http.StatusBadRequest, utils.NewErrorResponse(400, "Bad Request",
			"API ID is required"))
		return
	}

	// Extract optional deployment ID from query parameter
	deploymentID := c.Query("deploymentId")
	var deploymentIDPtr *string
	if deploymentID != "" {
		deploymentIDPtr = &deploymentID
	}

	// Parse and validate request body
	var notification dto.DeploymentNotification
	if err := c.ShouldBindJSON(&notification); err != nil {
		h.slogger.Warn("Invalid request body", "clientIP", clientIP, "error", err)
		c.JSON(http.StatusBadRequest, utils.NewErrorResponse(400, "Bad Request",
			"Invalid request body: "+err.Error()))
		return
	}

	// Create API and deployment using the service
	orgID := gateway.OrganizationID
	gatewayID := gateway.ID

	response, err := h.gatewayInternalService.CreateGatewayDeployment(
		apiID, orgID, gatewayID, notification, deploymentIDPtr)
	if err != nil {
		if errors.Is(err, constants.ErrInvalidInput) {
			c.JSON(http.StatusBadRequest, utils.NewErrorResponse(400, "Bad Request",
				"Invalid input data"))
			return
		}
		if errors.Is(err, constants.ErrGatewayNotFound) {
			c.JSON(http.StatusNotFound, utils.NewErrorResponse(404, "Not Found",
				"Gateway not found"))
			return
		}
		if errors.Is(err, constants.ErrAPINotFound) {
			c.JSON(http.StatusNotFound, utils.NewErrorResponse(404, "Not Found",
				"API not found"))
			return
		}
		h.slogger.Error("Failed to create gateway API deployment", "apiID", apiID, "gatewayID", gatewayID, "error", err)
		c.JSON(http.StatusInternalServerError, utils.NewErrorResponse(500, "Internal Server Error",
			"Failed to create API deployment"))
		return
	}

	h.slogger.Info("Successfully created gateway API deployment", "apiID", apiID, "gatewayID", gatewayID, "created", response.Created)

	// Return success response
	c.JSON(http.StatusCreated, map[string]interface{}{
		"message": response.Message,
	})
}

// GetLLMProvider handles GET /api/internal/v1/llm-providers/:providerId
func (h *GatewayInternalAPIHandler) GetLLMProvider(c *gin.Context) {
	// Extract client IP for rate limiting
	clientIP := c.ClientIP()

	// Extract and validate API key from header
	apiKey := c.GetHeader("api-key")
	if apiKey == "" {
		h.slogger.Warn("Unauthorized access attempt - Missing API key", "clientIP", clientIP)
		c.JSON(http.StatusUnauthorized, utils.NewErrorResponse(401, "Unauthorized",
			"API key is required. Provide 'api-key' header."))
		return
	}

	// Authenticate gateway using API key
	gateway, err := h.gatewayService.VerifyToken(apiKey)
	if err != nil {
		h.slogger.Warn("Authentication failed", "clientIP", clientIP, "error", err)
		c.JSON(http.StatusUnauthorized, utils.NewErrorResponse(401, "Unauthorized",
			"Invalid or expired API key"))
		return
	}

	orgID := gateway.OrganizationID
	gatewayID := gateway.ID
	providerID := c.Param("providerId")
	if providerID == "" {
		c.JSON(http.StatusBadRequest, utils.NewErrorResponse(400, "Bad Request",
			"Provider ID is required"))
		return
	}

	provider, err := h.gatewayInternalService.GetActiveLLMProviderDeploymentByGateway(providerID, orgID, gatewayID)
	if err != nil {
		if errors.Is(err, constants.ErrDeploymentNotActive) {
			c.JSON(http.StatusNotFound, utils.NewErrorResponse(404, "Not Found",
				"No active deployment found for this LLM provider on this gateway"))
			return
		}
		if errors.Is(err, constants.ErrLLMProviderNotFound) {
			c.JSON(http.StatusNotFound, utils.NewErrorResponse(404, "Not Found",
				"LLM provider not found"))
			return
		}
		c.JSON(http.StatusInternalServerError, utils.NewErrorResponse(500, "Internal Server Error",
			"Failed to get LLM provider"))
		return
	}

	// Create ZIP file from LLM provider YAML file
	zipData, err := utils.CreateLLMProviderYamlZip(provider)
	if err != nil {
		h.slogger.Error("Failed to create ZIP file", "providerID", providerID, "error", err)
		c.JSON(http.StatusInternalServerError, utils.NewErrorResponse(500, "Internal Server Error",
			"Failed to create LLM provider package"))
		return
	}

	// Set headers for ZIP file download
	c.Header("Content-Type", "application/zip")
	c.Header("Content-Disposition", fmt.Sprintf("attachment; filename=\"llm-provider-%s.zip\"", providerID))
	c.Header("Content-Length", fmt.Sprintf("%d", len(zipData)))

	// Return ZIP file
	c.Data(http.StatusOK, "application/zip", zipData)
}

// GetLLMProxy handles GET /api/internal/v1/llm-proxies/:proxyId
func (h *GatewayInternalAPIHandler) GetLLMProxy(c *gin.Context) {
	// Extract client IP for rate limiting
	clientIP := c.ClientIP()

	// Extract and validate API key from header
	apiKey := c.GetHeader("api-key")
	if apiKey == "" {
		h.slogger.Warn("Unauthorized access attempt - Missing API key", "clientIP", clientIP)
		c.JSON(http.StatusUnauthorized, utils.NewErrorResponse(401, "Unauthorized",
			"API key is required. Provide 'api-key' header."))
		return
	}

	// Authenticate gateway using API key
	gateway, err := h.gatewayService.VerifyToken(apiKey)
	if err != nil {
		h.slogger.Warn("Authentication failed", "clientIP", clientIP, "error", err)
		c.JSON(http.StatusUnauthorized, utils.NewErrorResponse(401, "Unauthorized",
			"Invalid or expired API key"))
		return
	}

	orgID := gateway.OrganizationID
	gatewayID := gateway.ID
	proxyID := c.Param("proxyId")
	if proxyID == "" {
		c.JSON(http.StatusBadRequest, utils.NewErrorResponse(400, "Bad Request",
			"Proxy ID is required"))
		return
	}

	proxy, err := h.gatewayInternalService.GetActiveLLMProxyDeploymentByGateway(proxyID, orgID, gatewayID)
	if err != nil {
		if errors.Is(err, constants.ErrDeploymentNotActive) {
			c.JSON(http.StatusNotFound, utils.NewErrorResponse(404, "Not Found",
				"No active deployment found for this LLM proxy on this gateway"))
			return
		}
		if errors.Is(err, constants.ErrLLMProxyNotFound) {
			c.JSON(http.StatusNotFound, utils.NewErrorResponse(404, "Not Found",
				"LLM proxy not found"))
			return
		}
		c.JSON(http.StatusInternalServerError, utils.NewErrorResponse(500, "Internal Server Error",
			"Failed to get LLM proxy"))
		return
	}

	// Create ZIP file from LLM proxy YAML file
	zipData, err := utils.CreateLLMProxyYamlZip(proxy)
	if err != nil {
		h.slogger.Error("Failed to create ZIP file", "proxyID", proxyID, "error", err)
		c.JSON(http.StatusInternalServerError, utils.NewErrorResponse(500, "Internal Server Error",
			"Failed to create LLM proxy package"))
		return
	}

	// Set headers for ZIP file download
	c.Header("Content-Type", "application/zip")
	c.Header("Content-Disposition", fmt.Sprintf("attachment; filename=\"llm-proxy-%s.zip\"", proxyID))
	c.Header("Content-Length", fmt.Sprintf("%d", len(zipData)))

	// Return ZIP file
	c.Data(http.StatusOK, "application/zip", zipData)
}

// ReceiveGatewayManifest handles POST /api/internal/v1/gateways/:gatewayId/manifest
// Called by the gateway controller to post back its installed custom policy manifest.
func (h *GatewayInternalAPIHandler) ReceiveGatewayManifest(c *gin.Context) {
	clientIP := c.ClientIP()

	apiKey := c.GetHeader("api-key")
	if apiKey == "" {
		h.slogger.Warn("Unauthorized access attempt - Missing API key", "clientIP", clientIP)
		c.JSON(http.StatusUnauthorized, utils.NewErrorResponse(401, "Unauthorized",
			"API key is required. Provide 'api-key' header."))
		return
	}

	gateway, err := h.gatewayService.VerifyToken(apiKey)
	if err != nil {
		h.slogger.Warn("Authentication failed", "clientIP", clientIP, "error", err)
		c.JSON(http.StatusUnauthorized, utils.NewErrorResponse(401, "Unauthorized",
			"Invalid or expired API key"))
		return
	}

	gatewayID := c.Param("gatewayId")
	if gatewayID == "" || gatewayID != gateway.ID {
		c.JSON(http.StatusForbidden, utils.NewErrorResponse(403, "Forbidden",
			"Gateway ID mismatch"))
		return
	}

	var body struct {
		Policies []service.GatewayPolicyInput `json:"policies"`
	}
	if err := c.ShouldBindJSON(&body); err != nil {
		c.JSON(http.StatusBadRequest, utils.NewErrorResponse(400, "Bad Request", err.Error()))
		return
	}

	if err := h.gatewayService.ReceiveGatewayManifest(gatewayID, body.Policies); err != nil {
		h.slogger.Error("Failed to store gateway manifest", "gatewayID", gatewayID, "error", err)
		c.JSON(http.StatusInternalServerError, utils.NewErrorResponse(500, "Internal Server Error",
			"Failed to store gateway manifest"))
		return
	}

	c.Status(http.StatusNoContent)
}

func (h *GatewayInternalAPIHandler) RegisterRoutes(r *gin.Engine) {
	orgGroup := r.Group("/api/internal/v1/apis")
	{
		orgGroup.GET("", h.GetAPIsByOrganization)
		orgGroup.GET("/:apiId", h.GetAPI)
		orgGroup.POST("/:apiId/gateway-deployments", h.CreateGatewayDeployment)
	}

	llmGroup := r.Group("/api/internal/v1/llm-providers")
	{
		llmGroup.GET("/:providerId", h.GetLLMProvider)
	}

	llmProxyGroup := r.Group("/api/internal/v1/llm-proxies")
	{
		llmProxyGroup.GET("/:proxyId", h.GetLLMProxy)
	}

	gatewayGroup := r.Group("/api/internal/v1/gateways")
	{
		gatewayGroup.POST("/:gatewayId/manifest", h.ReceiveGatewayManifest)
	}
}