# platform-api/src/resources/openapi.yaml
openapi: 3.1.1
info:
  title: WSO2 API Platform - Platform API
  description: |
    This document specifies a **RESTful API** for WSO2 **API Platform** - **Platform API**.
    
    ## Authentication
    
    All API endpoints (except organization registration) require authentication using JWT access tokens 
    generated by Thunder STS (Security Token Service).
    
    ### JWT Token Requirements
    
    - **Token Format**: Bearer token in Authorization header
    - **Required Claim**: `organization` - UUID of the user's organization
    - **Token Validation**: In development mode, signature validation can be skipped using `JWT_SKIP_VALIDATION=true`
    
    ### Organization Isolation
    
    All API operations are automatically scoped to the organization specified in the JWT token's 
    `organization` claim. Organization IDs are no longer required in request bodies or query parameters.

  version: 1.0.0
  contact:
    name: WSO2 API Platform Team
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://localhost:9243/api/v1
    description: Development server (HTTPS)
  - url: https://api.platform.com/api/v1
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /rest-apis/{apiId}/gateways:
    get:
      summary: Get gateways for REST API
      description: |
        Retrieves all gateways associated with the specified API, including deployment details.
        Returns gateway information along with association timestamps and deployment status.
        Access is validated against the organization in the JWT token.
      operationId: GetRESTAPIGateways
      tags:
        - REST APIs
        - Gateways
      parameters:
        - $ref: '#/components/parameters/apiId'
      responses:
        '200':
          description: List of gateways associated with the API, including deployment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RESTAPIGatewayListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Add gateways for REST API
      description: |
        Associates gateways to the specified API. If gateways are already associated,
        updates the association timestamp. Returns all gateways associated with the API
        including deployment details. Access is validated against the organization 
        in the JWT token.
      operationId: AddGatewaysToAPI
      tags:
        - REST APIs
        - Gateways
      parameters:
        - $ref: '#/components/parameters/apiId'
      requestBody:
        description: List of gateways to associate with the API
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/AddGatewayToRESTAPIRequest'
      responses:
        '200':
          description: List of all gateways associated with the API, including deployment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RESTAPIGatewayListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /gateways:
    post:
      summary: Register a new gateway
      description: |
        Creates a new gateway within the organization specified in the JWT token.
        Organization ID is automatically extracted from the token and does not need to be provided.
      operationId: CreateGateway
      tags:
        - Gateways
      requestBody:
        description: Gateway registration details
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGatewayRequest'
      responses:
        '201':
          description: Gateway registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Gateway name already exists within organization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalServerError'

    get:
      summary: List all gateways
      description: |
        Retrieves a list of all registered gateways for the organization specified in the JWT token.
        Organization ID is automatically extracted from the token.
      operationId: ListGateways
      tags:
        - Gateways
      responses:
        '200':
          description: Gateways retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /gateways/{gatewayId}:
    get:
      summary: Get gateway by UUID
      description: |
        Retrieves a specific gateway by its UUID. Access is validated against the organization 
        in the JWT token.
      operationId: GetGateway
      tags:
        - Gateways
      parameters:
        - $ref: '#/components/parameters/GatewayID'
      responses:
        '200':
          description: Gateway retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      summary: Update gateway
      description: |
        Updates an existing gateway's mutable fields (description).
        Access is validated against the organization in the JWT token.
      operationId: UpdateGateway
      tags:
        - Gateways
      parameters:
        - $ref: '#/components/parameters/GatewayID'
      requestBody:
        description: Gateway object that needs to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGatewayRequest'
      responses:
        '200':
          description: Gateway updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Delete gateway
      description: |
        Permanently deletes a gateway and all associated tokens (CASCADE).
        Deletion is blocked if the gateway has active API deployments or WebSocket connections.
        Access is validated against the organization in the JWT token.
      operationId: DeleteGateway
      tags:
        - Gateways
      parameters:
        - $ref: '#/components/parameters/GatewayID'
      responses:
        '204':
          description: Gateway deleted successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Conflict. Gateway has active deployments or connections
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                activeDeployments:
                  value:
                    code: 409
                    message: Conflict
                    description: Cannot delete gateway - 3 active API deployment(s) exist
                activeConnections:
                  value:
                    code: 409
                    message: Conflict
                    description: Cannot delete gateway - 2 active connection(s) exist
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /status/gateways:
    get:
      summary: Get gateway status
      description: |
        Retrieves lightweight status information for gateways. This endpoint is optimized for frequent polling 
        by management portals to monitor gateway connection status. If gatewayId query parameter is provided, 
        returns status for that specific gateway only. Otherwise returns status for all gateways in the organization.
      operationId: GetGatewayStatus
      tags:
        - Gateways
      parameters:
        - $ref: '#/components/parameters/gatewayId-Q'
      responses:
        '200':
          description: Gateway status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayStatusListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Gateway not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /llm-providers/{id}:
    get:
      summary: Get LLM provider by identifier
      description: Retrieve the complete configuration for a specific LLM provider.
      operationId: getLLMProvider
      tags:
        - LLM Providers
      parameters:
        - name: id
          in: path
          required: true
          description: Unique identifier of the LLM provider
          schema:
            type: string
      responses:
        '200':
          description: LLM provider details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LLMProvider'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      summary: Update an existing LLM provider
      description: Update the configuration of an existing LLM provider.
      operationId: updateLLMProvider
      tags:
        - LLM Providers
      parameters:
        - name: id
          in: path
          required: true
          description: Unique identifier of the LLM provider
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LLMProvider'
      responses:
        '200':
          description: LLM provider updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LLMProvider'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Delete an LLM provider
      description: Remove an LLM provider.
      operationId: deleteLLMProvider
      tags:
        - LLM Providers
      parameters:
        - name: id
          in: path
          required: true
          description: Unique identifier of the LLM provider
          schema:
            type: string
      responses:
        '204':
          description: LLM provider deleted successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /llm-proxies/{id}:
    get:
      summary: Get LLM proxy by unique identifier
      description: Retrieve the complete configuration for a specific LLM proxy.
      operationId: getLLMProxy
      tags:
        - LLM Proxies
      parameters:
        - name: id
          in: path
          required: true
          description: Unique identifier of the LLM proxy
          schema:
            type: string
      responses:
        '200':
          description: LLM proxy details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LLMProxy'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      summary: Update an existing LLM proxy
      description: Update the configuration of an existing LLM proxy.
      operationId: updateLLMProxy
      tags:
        - LLM Proxies
      parameters:
        - name: id
          in: path
          required: true
          description: Unique identifier of the LLM proxy
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LLMProxy'
      responses:
        '200':
          description: LLM proxy updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LLMProxy'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Delete an LLM proxy
      description: Remove an LLM proxy.
      operationId: deleteLLMProxy
      tags:
        - LLM Proxies
      parameters:
        - name: id
          in: path
          required: true
          description: Unique identifier of the LLM proxy
          schema:
            type: string
      responses:
        '204':
          description: LLM proxy deleted successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token issued by Thunder STS. The token must contain an `organization` claim 
        that identifies the user's organization. All API operations are automatically scoped 
        to this organization.
        
        Example: `Authorization: Bearer eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9...`
  schemas:
...
    AddGatewayToRESTAPIRequest:
      title: AddGatewayToAPIRequest object with basic gateway details
      type: object
      required:
        - gatewayId
      properties:
        gatewayId:
          type: string
          format: uuid
          example: "987e6543-e21b-45d3-a789-426614174999"
    RESTAPIGatewayListResponse:
      type: object
      required:
        - count
        - list
        - pagination
      properties:
        count:
          type: integer
          description: Number of gateways in current response
          example: 3
        list:
          type: array
          description: List of gateways associated with the API, including deployment details when deployed
          items:
            $ref: '#/components/schemas/RESTAPIGatewayResponse'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    CreateGatewayRequest:
      description: |
        Request body for creating a gateway. Organization ID is automatically extracted 
        from the JWT token and does not need to be provided.
      type: object
      required:
        - name
        - displayName
        - vhost
        - functionalityType
      properties:
        name:
          type: string
          pattern: '^[a-z0-9-]+$'
          minLength: 3
          maxLength: 64
          description: URL-friendly gateway identifier (lowercase alphanumeric with hyphens, unique per organization)
          example: "prod-gateway-01"
        displayName:
          type: string
          minLength: 1
          maxLength: 128
          description: Human-readable gateway name
          example: "Production Gateway 01"
        description:
          type: string
          description: Description of the gateway
          example: "Production gateway for handling API traffic"
        vhost:
          type: string
          description: Virtual host (domain name) for the gateway
          maxLength: 255
          minLength: 1
          # hostname regex as per RFC 1123 (http://tools.ietf.org/html/rfc1123) and appended *
          pattern: '^(([a-zA-Z0-9]|[a-zA-Z0-9][a-zA-Z0-9\-]*[a-zA-Z0-9])\.)*([A-Za-z0-9]|[A-Za-z0-9][A-Za-z0-9\-]*[A-Za-z0-9])$'
          example: mg.wso2.com
        isCritical:
          type: boolean
          description: Whether the gateway is critical for production
          default: false
          example: true
        functionalityType:
          type: string
          enum: [regular, ai, event]
          description: Type of gateway functionality
          default: regular
          example: "regular"
        properties:
          type: object
          additionalProperties: true
          description: Custom key-value properties for the gateway
          example: {"region": "us-west", "tier": "premium"}
...
  responses:
    Unauthorized:
      description: Unauthorized. Authentication credentials are missing or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 401
            message: Unauthorized
            description: Authorization header is required or token is invalid
    BadRequest:
      description: Bad Request. Invalid request or validation error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 400
            message: Bad Request
            description: Invalid request or validation error
    NotFound:
      description: Not Found. The specified resource does not exist.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 404
            message: Not Found
            description: The specified resource does not exist
...
    InternalServerError:
      description: Internal Server Error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 500
            message: Internal Server Error
            description: The server encountered an internal error. Please contact
              administrator.
  parameters:
...
    apiId:
      name: apiId
      in: path
      description: |
        **API ID** consisting of the **handle** (unique identifier) of the API.
      required: true
      schema:
        type: string
        example: my-rest-api-handle
...